+++ tests/PSRule.Rules.Azure.Tests/FunctionTests.cs
@@ -196,6 +196,19 @@ namespace PSRule.Rules.Azure
 
         [Fact]
         [Trait(TRAIT, TRAIT_ARRAY)]
+        public void Null()
+        {
+            var context = GetContext();
+
+            var actual1 = Functions.Null(context, null);
+            var actual2 = Functions.Null(context, new object[] { });
+
+            Assert.Null(actual1);
+            Assert.Null(actual2);
+        }
+
+        [Fact]
+        [Trait(TRAIT, TRAIT_ARRAY)]
         public void Last()
         {
             var context = GetContext();
@@ -810,6 +823,20 @@ namespace PSRule.Rules.Azure
 
         [Fact]
         [Trait(TRAIT, TRAIT_LOGICAL)]
+        public void False()
+        {
+            var context = GetContext();
+
+            var actual1 = (bool)Functions.False(context, null);
+            var actual2 = (bool)Functions.False(context, new object[] { });
+            Assert.False(actual1);
+            Assert.False(actual2);
+
+            Assert.Throws<ExpressionArgumentException>(() => Functions.False(context, new object[] { "true" }));
+        }
+
+        [Fact]
+        [Trait(TRAIT, TRAIT_LOGICAL)]
         public void If()
         {
             var context = GetContext();
@@ -844,6 +871,20 @@ namespace PSRule.Rules.Azure
             Assert.True(actual1);
             Assert.True(actual2);
             Assert.False(actual3);
+        }
+
+        [Fact]
+        [Trait(TRAIT, TRAIT_LOGICAL)]
+        public void True()
+        {
+            var context = GetContext();
+
+            var actual1 = (bool)Functions.True(context, null);
+            var actual2 = (bool)Functions.True(context, new object[] { });
+            Assert.True(actual1);
+            Assert.True(actual2);
+
+            Assert.Throws<ExpressionArgumentException>(() => Functions.True(context, new object[] { "true" }));
         }
 
         #endregion Logical
@@ -896,8 +937,6 @@ namespace PSRule.Rules.Azure
             var actual2 = (float)Functions.Float(context, new object[] { "3.0" });
             Assert.Equal(3.0f, actual2);
 
-            // String
-
             Assert.Throws<ExpressionArgumentException>(() => Functions.Float(context, null));
             Assert.Throws<ExpressionArgumentException>(() => Functions.Float(context, new object[] { }));
             Assert.Throws<FormatException>(() => Functions.Float(context, new object[] { "one" }));
@@ -916,8 +955,6 @@ namespace PSRule.Rules.Azure
             // String
             var actual2 = (int)Functions.Int(context, new object[] { "4" });
             Assert.Equal(4, actual2);
-
-            // String
 
             Assert.Throws<ExpressionArgumentException>(() => Functions.Int(context, null));
             Assert.Throws<ExpressionArgumentException>(() => Functions.Int(context, new object[] { }));